portable zeek
This patch modifies necessary files to make zeek and zeek config
"portable". This means that paths reported by zeek will paths relative
to the zeek or zeek-config executable file instead of paths being
hard-coded to those used at install time.
This makes it so that zeek installation tree can be copied to a
different install prefix (as is done by dependency managers like vcpkg
or conan).